pyproject.toml
[project]
name = "my-agent-app"
version = "0.1.0"
description = "Agente inteligente com FastAPI + LangChain + MCP"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.135.1",
    "jinja2>=3.1.6",
    "langchain[anthropic]>=1.2.12",
    "langchain-mcp-adapters>=0.1.0",
    "sqlalchemy>=2.0.48",
    "asyncpg>=0.30.0",
    "alembic>=1.15.0",
    "uvicorn>=0.41.0",
    "python-dotenv>=1.0.0",
    "markdown>=3.7",
    "httpx>=0.28.0",
]

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/my_agent_app"]
